+++ data/serializer/src/main/java/tech/pegasys/teku/api/response/v1/node/Meta.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 ConsenSys AG.
+ * Copyright 2021 ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -15,18 +15,24 @@ package tech.pegasys.teku.api.response.v1.node;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
+import com.google.common.base.MoreObjects;
+import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Objects;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public class PeersResponse {
-  public final List<Peer> data;
-  public final Meta meta;
+public class Meta {
+  @JsonProperty("count")
+  @Schema(type = "string", description = "Total number of items", example = "1")
+  public final UInt64 count;
 
   @JsonCreator
-  public PeersResponse(
-      @JsonProperty("data") final List<Peer> data, @JsonProperty("meta") final Meta meta) {
-    this.data = data;
-    this.meta = meta;
+  public Meta(@JsonProperty("count") final UInt64 count) {
+    this.count = count;
+  }
+
+  @Override
+  public String toString() {
+    return MoreObjects.toStringHelper(this).add("count", count).toString();
   }
 
   @Override
@@ -37,12 +43,12 @@ public class PeersResponse {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    final PeersResponse that = (PeersResponse) o;
-    return Objects.equals(data, that.data) && Objects.equals(meta, that.meta);
+    Meta meta = (Meta) o;
+    return Objects.equals(count, meta.count);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(data, meta);
+    return Objects.hash(count);
   }
 }